order total integer to float

+++ db/migrate/20220418024140_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[7.0]
   def change
     create_table :orders do |t|
       t.integer :customer_id
-      t.integer :total
+      t.float :total
       t.date :order_date
       t.string :status
